logs_generator/requirements.txt
faker>=18.0.0
pyyaml>=6.0
numpy>=1.24.0
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.0.0
